[ docs ] 메인으로 index 변경
